+++ src/config.js
@@ -1,4 +1,4 @@
-const BASE_URL = 'http://10.58.7.170:3000';
+const BASE_URL = 'http://10.58.0.224:3000';
 const API = {
   home: `${BASE_URL}/product/main`,
   product: `${BASE_URL}/product/all`,
