+++ src/pages/ListPage/ListPage.scss
@@ -1,6 +1,6 @@
 .listPage {
   .contentTitle {
-    margin: 60px 0 48px;
+    margin: 120px 0 48px;
     font-size: 24px;
     text-align: center;
     line-height: 54px;
